+++ src/autopas/utils/WrapMPI.h
@@ -715,7 +715,8 @@ inline int AutoPas_MPI_Request_free(AutoPas_MPI_Request *request) {
 inline int AutoPas_MPI_Allgather(void *buffer_send, int count_send, AutoPas_MPI_Datatype datatype_send,
                                  void *buffer_recv, int count_recv, AutoPas_MPI_Datatype datatype_recv,
                                  AutoPas_MPI_Comm comm) {
-  for (int i = 0; i < (count_recv / count_send); i++)
+  for (long i = 0; i < (count_recv / count_send); i++)
+    // offsets from pointers are of type ptrdiff_t which is an alias for long. Hence, i should be long.
     memcpy(static_cast<char *>(buffer_recv) + (i * count_send * sizeof(datatype_send)), buffer_send,
            count_send * sizeof(datatype_send));
   return AUTOPAS_MPI_SUCCESS;
